micro refactoring and docs

+++ Business/Interfaces/IStorageService.cs
@@ -8,10 +8,10 @@ namespace CadwiseAutomaticTellerMachine.Business.Interfaces
         Task<int> GetStorageBalance();
         Task<List<BanknoteQuantityDto>> GetBanknoteQuantity();
 
-        Task<List<BanknoteQuantityDto>> WithdrawMoneyBig(int moneyRequest);
-        Task<List<BanknoteQuantityDto>> WithdrawMoneySmall(int moneyRequest);
-        int ProcessWithdraw(List<BanknoteQuantityDto> denominations, int request, List<BanknoteQuantityDto> result);
-        Task ValidateMoneyRequest(int moneyRequest);
+        Task<List<BanknoteQuantityDto>> WithdrawMoneyBig(long moneyRequest);
+        Task<List<BanknoteQuantityDto>> WithdrawMoneySmall(long moneyRequest);
+        long ProcessWithdraw(List<BanknoteQuantityDto> denominations, long request, List<BanknoteQuantityDto> result);
+        Task ValidateMoneyRequest(long moneyRequest);
         Task IncreaseBanknoteQuantity(List<BanknoteQuantityDto> banknotesQuantity);
     }
 }
